+++ lib/gui/app/components/settings/settings.tsx
@@ -19,7 +19,8 @@ import * as _ from 'lodash';
 import * as React from 'react';
 import { Box, Checkbox, Flex, Txt } from 'rendition';
 
-import { version, packageType } from '../../../../../package.json';
+import { version } from '../../../../../package.json';
+import { isUpdateSupported } from '../../../../shared/update-support';
 import * as settings from '../../models/settings';
 import { open as openExternal } from '../../os/open-external/services/open-external';
 import { Modal } from '../../styled-components';
@@ -42,7 +43,7 @@ async function getSettingsList(): Promise<Setting[]> {
 			label: i18next.t('settings.trimExtPartitions'),
 		},
 	];
-	if (['appimage', 'nsis', 'dmg'].includes(packageType)) {
+	if (isUpdateSupported()) {
 		list.push({
 			name: 'updatesEnabled',
 			label: i18next.t('settings.autoUpdate'),